+++ HISTORY.md
@@ -2,6 +2,7 @@ unreleased
 ==========
 
   * Attempt to combine multiple ranges into single range
+  * Ignore non-byte `Range` headers
   * deps: http-errors@~1.5.0
     - Add `HttpError` export, for `err instanceof createError.HttpError`
     - Support new code `421 Misdirected Request`